Build Pynq Images with NixOS
This uses NixOS to cross-compile full NixOS sd images ready to dd to an
sd-card, as well as system closures to be used with nix-copy-closure for
incremental updates.

The old "build Linux kernel with Nix, but use ArchlinuxARM for the
userland approach is removed, as well as the pynqBootFS helper.